Fixed XML output
The main alteration involved in this fix was changing the transformation
matrix member of the RotateAction subclass from a distinctly multidimensional
array to a jagged array. (basically, replaced the "[,]" with a "[][]".

I've seen no appreciable changes in how the program operates as a result of
this change in array formatting, however I am newest on the project. So, if
you start noticing weird indexing problems, this may be the reason.