Resolved missing file issue. However, XML input from the table webpage required.
Once the program waits for a cue to continue, one must open up the
"partTable.html" file, open up the XML file named "parts_properties.xml" via
the web page, then fill out the missing information in the table. From there,
one must render the XML representation of the new table and save the resulting
file, "parts_properties2.xml", to the same folder as the Assembly Planner
executable. Afterwards, navigate back to the console of the main application
and press enter to have the program read in the xml and complete the job.